+++ src/app/api/forgotPassword/route.ts
@@ -1,4 +1,4 @@
-import clientPromise from '@/app/lib/mongodb';
+import clientPromise from '@/lib/mongodb';
 import { NextRequest, NextResponse } from 'next/server';
 import otpGenerator from 'otp-generator';
