+++ src/app/api/signup/route.ts
@@ -1,4 +1,4 @@
-import clientPromise from '@/app/lib/mongodb';
+import clientPromise from '@/lib/mongodb';
 import bcrypt from 'bcrypt';
 import { NextResponse } from 'next/server';
